fix:: activity에서 fragment로 옮김.
