+++ app/src/main/AndroidManifest.xml
@@ -12,11 +12,21 @@
         android:theme="@style/Theme.SharedReceiptsApplication"
         tools:targetApi="31">
         <activity
+            android:name=".NavitaionActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+        <activity
             android:name=".ToggleBtActivity"
             android:exported="false" />
         <activity
-            android:name=".MypageActivity"
-            android:exported="false" />
+            android:name=".FavoriteActivity"
+            android:exported="true">
+        </activity>
         <activity
             android:name=".WrittingScript2Activity"
             android:exported="false" />
@@ -25,8 +35,7 @@
             android:exported="false" />
         <activity
             android:name=".DropOutActivity"
-            android:exported="true">
-        </activity>
+            android:exported="true"></activity>
         <activity
             android:name=".ProflieSettingActivity"
             android:exported="true" />
@@ -45,11 +54,6 @@
         <activity
             android:name=".MainActivity"
             android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
     </application>
