--- app/src/main/java/com/example/shared_receipts_application/MypageActivity.kt
@@ -1,12 +1,0 @@
-package com.example.shared_receipts_application
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MypageActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_mypage)
-
-    }
-}
