+++ app/src/main/res/layout/activity_favorite.xml
@@ -5,27 +5,4 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content">
 
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="56dp"
-        android:text="찜"
-        android:textSize="20dp"
-        android:textStyle="bold"
-        android:textColor="#000"
-        android:gravity="center"
-        android:layout_gravity="center"/>
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="10dp"
-        android:background="@drawable/drop_shadow" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/favorite_"
-        android:layout_width="365dp"
-        android:layout_height="545dp"
-        android:layout_marginTop="20dp"
-        android:layout_marginLeft="23dp">
-
-    </androidx.recyclerview.widget.RecyclerView>
-
 </LinearLayout>
