+++ app/src/main/res/layout/activity_main.xml
@@ -56,22 +56,5 @@
         android:layout_height="wrap_content"
         android:layout_marginLeft="23dp">
     </androidx.recyclerview.widget.RecyclerView>
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/nav_host"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:name="androidx.navigation.fragment.NavHostFragment"
-        app:defaultNavHost="true"
-        app:navGraph="@navigation/nav_gragh"
-        />
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/bottom_nav"
-        android:layout_width="match_parent"
-        android:layout_height="56dp"
-        android:layout_gravity="bottom"
-        app:labelVisibilityMode="labeled"
-        app:itemBackground="@color/white"
-        app:menu="@menu/button_menu"
-        app:layout_constraintBottom_toBottomOf="parent"
-        />
+
 </LinearLayout>
