+++ app/src/main/res/layout/fragment_heart.xml
@@ -1,14 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/heartFragment"
     tools:context="com.example.shared_receipts_application.fragments.HeartFragment">
-    <TextView
+    <LinearLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        android:orientation="vertical"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:layout_centerInParent="true"
-        android:text="Heart"/>
+        android:layout_height="wrap_content">
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="56dp"
+            android:text="찜"
+            android:textSize="20dp"
+            android:textStyle="bold"
+            android:textColor="#000"
+            android:gravity="center"
+            android:layout_gravity="center"/>
+        <View
+            android:layout_width="match_parent"
+            android:layout_height="10dp"
+            android:background="@drawable/drop_shadow" />
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/favorite_"
+            android:layout_width="365dp"
+            android:layout_height="545dp"
+            android:layout_marginTop="20dp"
+            android:layout_marginLeft="23dp">
+
+        </androidx.recyclerview.widget.RecyclerView>
+    </LinearLayout>
 </RelativeLayout>
